apply a brand new function to perform requests, remove index.ts file from lib folder to split server-side and client-side files
